#375 EditControllerの機能をUsersControllerへ集約する(修正)
