+++ routes/web.php
@@ -35,9 +35,9 @@ Route::group(['middleware' => 'auth'], function () {
     Route::delete('{id}', 'PostsController@destroy')->name('post.delete');
     //ユーザー編集
     Route::prefix('users/{id}')->group(function () {
-        Route::get('edit', 'EditUserController@edit')->name('edit'); 
-        Route::put('', 'EditUserController@update')->name('update');
-        Route::delete('', 'EditUserController@destroy')->name('user.delete');
+        Route::get('edit', 'UsersController@edit')->name('edit'); 
+        Route::put('', 'UsersController@update')->name('update');
+        Route::delete('', 'UsersController@destroy')->name('user.delete');
     //フォロー機能
         Route::post('follow', 'FollowersController@store')->name('follow');
         Route::delete('unfollow', 'FollowersController@destroy')->name('unfollow');
